ml-dev/projects/prop_model/requirements.txt
google-api-core==2.19.1
google-auth==2.34.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
altair<5
numpy==1.24.4
pandas==2.0.3
pyarrow==17.0.0
pybigquery==0.10.2
pydantic==2.8.2
pydantic_core==2.20.1
streamlit==1.24.0
tiktoken==0.7.0
toml==0.10.2
scikit-learn==1.3.2
matplotlib
seaborn
sklearn
xgboost==1.6.1
category_encoders
db-dtypes
shap
streamlit_shap
